handle UnionT ~> NotT flows properly
Summary:
`NotT` assumes that all lower bounds that are not concrete types like Booleans or Numbers are truthy, which is not true for unions, which might contain both truthy and falsy components. This splits the union to handle this case properly.

Previously, the following would spuriously error:
```
function triv () : true {
  declare var x : false | false;
  return !x;
}
```

Reviewed By: panagosg7

Differential Revision: D20953333

fbshipit-source-id: 4e2a8a42308ee41e0e8e3f0de2e69323e7c3de3a