Intern and print internal comments in empty blocks
Summary:
There is one more type of location that comments can appear where comments are not being attached to the AST, and are lost when reprinting. This is comments that are "internal", in this case meaning they are inside an empty node of some form.

One such place where comments can appear is inside an empty block. Luckily `Syntax.t` in the flow AST is already parameterized by the type of internal comments, so we can simply define the comment type for blocks to have an internal type of `Loc.t Comment.t list`.

Blocks can appear in block statements, function block bodies, and declare module statement. Let's make sure that each of these scenarios is implemented and tested.

There are quite a few other locations where internal comments can appear (e.g. inside empty object types, param lists, type param lists, etc.). For now let's just add internal comments in empty blocks, and create all the infrastructure and utilities for attaching other internal comment types to simplify their implementation. We should parameterize the utilities assuming that internal comments will be a list of comments, as this will be the case for most of the other nodes with internal comments.

Estree does not yet have a place to put internal comments, so let's report them as trailing comments. One can still tell they are internal from the reported types by inspecting the locs.

Reviewed By: mroch

Differential Revision: D21058356

fbshipit-source-id: b91806b1c9482e4db0efffa00ef48a6be381636f