Add regression test for duplicated comments at JSX lex mode change boundary
Summary:
We previously had a bug where comments at the boundary of lex mode changes were being duplicated in the AST. For example,

```
<div />; // A
<div />; // B
```

was having the comment "A" attached to the AST twice, as the lex mode change was wiping which comments were eaten.

This has already been fixed, but let's add a regression test since apparently this wasn't being caught before.

Test for github issue #8352

Reviewed By: gkz

Differential Revision: D21100992

fbshipit-source-id: 6b8625daa64486cc320b6566de60263af937d5aa