[enums] When missing check of member in switch, point at discriminant rather than entire switch
Summary:
When doing an exhaustive check of an enum in a switch, and we miss checking one of the enum members, we need to point at something for the error.

Previously, we pointed at the entire switch statement, but this is not the best because if the switch statement is large, it highlights a very large section of code.

Instead, we can point to the discriminant of the switch statement. This still points the user to where they need to add the additional case, without highlighting the entire switch statement.

I checked out what Rust does for this error and they point at the discriminant (as prior art).

Reviewed By: dsainati1

Differential Revision: D20999018

fbshipit-source-id: 0f38b5865184be5169ffcf94bcd63dceb7b90319