[enums] Allow importing enums only as type
Summary: Previously, you got an error if you imported an enum only as a type. Like a class, an enum is both a value and a type so can be imported as either. It is useful to import only as a type if you are only using the type, and so don't want to introduce an unnecessary runtime dependency.

Reviewed By: panagosg7

Differential Revision: D21034687

fbshipit-source-id: a750e4f14132960daec6226baec62b6a17d01bc9