[flow Intern and print internal comments in empty objects
Summary:
There is one more type of location that comments can appear where comments are not being attached to the AST, and are lost when reprinting. This is comments that are "internal", in this case meaning they are inside an empty node of some form.

One such place where comments can appear is inside an empty object. Just like for internal block comments lets attach a list of internal comments to objects, object types, and object patterns (which need general comment attachment implemented as well). We can use the internal comment utilities for this.

Reviewed By: pieterv

Differential Revision: D21058667

fbshipit-source-id: 7281389469d248f1c4e37f44b9a26672b3ef3e8f