Intern and print internal comments in empty JSX expression containers
Summary:
There is one more type of location that comments can appear where comments are not being attached to the AST, and are lost when reprinting. This is comments that are "internal", in this case meaning they are inside an empty node of some form.

One common place where comments can appear is inside an empty JSX expression container. Just like for internal block comments lets attach a list of internal comments to JSX expression containers. We can use the internal comment utilities for this.

Reviewed By: pieterv

Differential Revision: D21067592

fbshipit-source-id: 29ab7e742c0d5e85fa93e9e63eb052331e15c4a5