[enums] Treat the end of file as a valid end of enum member
Summary:
A trailing comma is optional for the last enum member. If the next token is the end of the file, then we are the last member, so we should only expect `}`, not also `,`.

This way, if have `enum E {A`, we just error with "Unexpected end of input, expected the token `}`" rather than additionally also erroring with "Unexpected end of input, expected the token `,`"

Note that we will always have the subsequent "Unexpected end of input, expected the token `}`" error, this just remove the unnecessary additional error.

Reviewed By: Hans-Halverson

Differential Revision: D21148916

fbshipit-source-id: 4f0db300ad578375c6540480cb1e739d00d6df1e