[enums] Add custom error when incorrectly using `:` as separator for initializer
Summary:
The valid separator between an enum member name and initializer is `=`.

```
enum E {
  A = 1,
}
```
is valid
```
enum E {
  A: 1,
}
```
is not.

However, it is possible for a user to incorrectly use `:`. Two common cases might be:
- If a user is refactoring an object to be an enum (e.g. refactoring `const E = Object.freeze({A: 1});` to an enum declaration), then they may accidentally use `:` as the separator if they forget to update it.
- Or, if they are simply used to writing object literals, they may make the same mistake.

Previously, the error would just be "Unexpected token `:`".  Instead, add a custom error that explains to the user what the correct separator is. Additionally, recover parsing more gracefully, assuming the user meant `:` as the separator, and parsing the rest of the enum with this assumption.

Reviewed By: Hans-Halverson

Differential Revision: D21146700

fbshipit-source-id: 7d07b898283dc9ec75e23220d53a896622d711a5