[enums] Add custom error when incorrectly using `;` as member separator
Summary:
The valid separator between enum members is `,`.

```
enum E {
  A,
  B
}
```
is valid
```
enum E {
  A;
  B
}
```
is not.

However, it is possible for a user to incorrectly use `,` if they are used to how Hack declares enums (they use `;` as the member separator).

Previously, the error would just be "Unexpected token `;`".  Instead, add a custom error that explains to the user what the correct separator is. Additionally, recover parsing more gracefully, assuming the user meant `;` as the separator, and parsing the rest of the enum with this assumption.

Reviewed By: Hans-Halverson

Differential Revision: D21147839

fbshipit-source-id: a0beb445202909f559ec072a8d317987c485c61d