[parser] Treat the end of file as a valid end of object property
Summary:
A trailing comma is optional for the last object property. If the next token is the end of the file, then we are the last property, so we should only expect `}`, not also `,`.

This way, if have `{x: 1`, we just error with "Unexpected end of input, expected the token `}`" rather than additionally also erroring with "Unexpected end of input, expected the token `,`"

Note that we will always have the subsequent "Unexpected end of input, expected the token `}`" error, this just remove the unnecessary additional error.

Found this while dealing with a similar case in enums: D21148916

Additionally note that until my work in the summer of 2019 we did not report what token we were expecting. Before then, we would have had two "Unexpected end of input" errors, which would have been deduplicated and only one would have been shown. Thus this issue was not apparent until my work to improve parser error messages.

Reviewed By: Hans-Halverson

Differential Revision: D21163526

fbshipit-source-id: 80b87f8faddc3d337e744446ff843a2cfbc4e12e