[enums] Check that enum member name is valid when analyzing switch
Summary:
For enums, names starting with lowercase 'a' through 'z' are reserved (it is a parse error to define such a member) for the enum methods.

Previously when checking the form of the switch case, we didn't check for this (just that it was a member access). We still errored later, but it is less expensive to store an known incorrect exhaustive check than a possibly valid exhaustive check.

Note that there are two errors in this case: one saying that the values (function and enum) are not compatible, and the other one about the form. I thought about changing it to only display one of the erorrs, however after some thought I believe it is better to display both, as this gives the user more information and a better chance of fixing all the issues at once.

Reviewed By: dsainati1

Differential Revision: D21181490

fbshipit-source-id: e90e703d10bed95f344ff89e61bf8e1385e1dd9d