Preserve empty lines and internal comments when printing arrays
Summary:
There is one more type of location that comments can appear where comments are not being attached to the AST, and are lost when reprinting. This is comments that are "internal", in this case meaning they are inside an empty node of some form.

One such place where comments can appear is inside an array (or array pattern). Just like for internal block comments lets attach a list of internal comments to array AST nodes. We can use the internal comment utilities for this.

```
[/* Internal */]

[
  a,
  /* Internal */
]
```

We also should preserve blank lines between elements, like we do for objects, using the `list_with_newlines` utility.

However, to preserve newlines correctly we also need to know where holes are in the array, so that we can preserve blank lines between them or between them and adjacent non-hole elements. Unfortunately the current representation of array holes in the AST does not preserve any location information. Instead, array elements are typed as a list of (expression or spread) options, where a `None` option represents a hole.

Instead, let's represent elements as a list of a new element type where element's can be expression, spreads, or holes. The hole variant contains a single loc (the loc of the comma).

Reviewed By: mroch

Differential Revision: D21102768

fbshipit-source-id: 85112b3ea428d7206e8a715f49a965af89c0550d