[codemod] Add "repeat mode" for command
Summary:
Some codemods (including annotate-exports) need to be run repeatedly until
no more changes can be made. This diff extends the codemod command to
allow for this. The codemod runner is passed an `iteration` argument that
holds the number of iteration. When
* iteration=0, the runner does an *initialization* of the server (Types_js.init) using
the input files as files to be checked. This part remains the same with before.
* iteration>0, the runner *rechecks* the files that were changed in the previous
iteration (Types_js.recheck). The set of files changed is computed using the
result of the differ-printer.

Since this mode requires rechecking ASTs, it cannot be combined with the "dry-run"
option.

Recheck takes care of oldifying heaps related to checking. Differ heaps are cleared
manually before every new iteration.

The iteration number is also passed to the codemod context, as it is needed for
some codemods.

Reviewed By: samwgoldman

Differential Revision: D20912202

fbshipit-source-id: f1bd90e5abaa466a7870edecd0d158e847a0c957