Fix duplicate comment attachment for object methods named "get" or "set"
Summary:
We currently have a bug where leading comments are attached to two nodes in object methods whose name is "get" or "set". Comments are attached both to the "get" or "set" identifier, as well as the overall function expression.

This is easy to fix, we just need to not forward the leading comments further on when we realize the "get" or "set" is the name of the method instead of indicating a getter or setter, since in this case the leading comments will already be attached to the key.

Reviewed By: gkz

Differential Revision: D21237085

fbshipit-source-id: 46103577ba69c04f113413f7fb9e21aa5fc3dd50