[enums] Handle exhaustive check as use_t, improving union handling
Summary: Previously we were saving the potential exhaustive checks to analyze to the context from `statement.ml`, and analyzing in `merge_js.ml`. By moving the logic that checks that the discriminant is an enum by adding `EnumExhaustiveCheckT` and moving that logic to `flow_js.ml`, we can take advantage of the existing machinery that handles unions and intersections. Additionally, we no longer store in the context possible exhaustive checks on non-enums, and invalid ones (which we handle immediately).

Reviewed By: dsainati1

Differential Revision: D21259823

fbshipit-source-id: 340a3f0aa993e301e1b31ec1f8b52cfac617a32e