Visit all variance nodes and their attached comments in AST mapper
Summary:
The AST mapper was not previously visiting the attached comments for variance nodes. We need to make sure these are visited, as the comment bound finder is built on the AST mapper.

In addition there were a few locations where variance nodes were not being visited in the AST mapper, so I fixed those cases as well.

This issue was presenting itself  in the `js_layout_generator`, as  leading comments on variance nodes were not being counted when calculating comment bounds. This was resulting in the following reprinting behavior, because the `js_layout_generator` is trying to preserve a newline between the two properties since it doesn't realize the comment should be attached to the property via the variance node:

```
// Before
type T = {
  +a: any,
  // L
  +b: any,
};

// After reprinting
type T = {
  +a: any,

  // L
  +b: any,
};
```

Reviewed By: gkz

Differential Revision: D21218092

fbshipit-source-id: 74c941f0fdd703366e013321e96d3d50d88e577c