Add option to preserve formatting in js_layout_generator
Summary:
We would like to be able to set particular options when running the `js_layout_generator`. The first option we would like to add is whether we should try to preserve as much formatting as possible in the original text when printing a file with the `js_layout_generator`.

We can address this by adding a new options object that is passed down through the calls to `js_layout_generator`. At the moment this options object simply contains a single field,  `preserve_formatting`, which will be used to enable/disable certain changes to the output depending on if original formatting should be preserved. We can also add more fields to this in the future when we have the need, but until this point there was no way of setting up any configuration for the `js_layout_generator`. I then updated all the layout generator functions to take this `opts` object as the first named argument for consistency. The default opts are to set `preserve_formatting = false` unless the caller has explicitly set it to be true. I threaded up the opts up to the top level `Replacement_printer` functions, and then set the printer in codemods to preserve formatting. Since the default value is to not preserve formatting, any other users of the `js_layout_generator` are not affected.

Finally I made it so that if `preserve_formatting` is true the `js_layout_generator` does not modify any literals. Instead, it simply prints back the raw original values.

Reviewed By: pieterv

Differential Revision: D21186992

fbshipit-source-id: b95bdd16ce843f74c81c9418ac874c6c53d4f766