Store dependency_info instead of file_sigs in the saved state
Summary:
For the most part, we don't actually need all of the information in the file_sigs in order to initialize a server without parsing the entire codebase (the goal of saved state). We just need the dependency graph.

Previously, we derived the dependency graph from the stored file sigs when initializing from saved state. Now, we just store the dependency graph directly. This significantly cuts down on the size of the saved state (~25% on one major codebase) and also reduces the time spent initializing.

This also fixes the tests enabled in D21092066, because the dependency_info structure includes the sig_dependency_graph which is required for types-first.

There is one place where we expect file sigs to be populated where we don't already call `ensure_parsed`. I simply added an `ensure_parsed` call there. The set of files is relatively small, and they would get parsed later in the recheck anyway.

At some point in the future I'll evaluate whether storing just the forward dependency graph is better. It will likely reduce the size of the saved state, but it will likely also increase the time spent to initialize from saved state.

Reviewed By: samwgoldman

Differential Revision: D21213710

fbshipit-source-id: 0dc53c798f2c12b1d372dbacc84e579da42eebda