Avoid wasteful string duplication in saved state
Summary:
When OCaml marshals a data structure, it avoids marshaling the same element of it twice. Unfortunately, the process of normalizing paths to be relative to the project root causes a new allocation each time a path is normalized. Then, since the resulting values are no longer physically equal, they are each marshaled separately, causing the saved state to be unnecessarily large.

This change caches the results of normalizing a `File_key.t`, thereby avoiding the allocation of duplicate normalized strings. This isn't quite a typical interning scheme, but it's a bit more time-efficient to avoid normalizing the file keys and then interning them.

Reviewed By: samwgoldman

Differential Revision: D21233334

fbshipit-source-id: c832b14cf402fd1792fe11ebc492f3b56132db07