Don't use multiworkers when saving the saved state
Summary:
Removing the multiworkers unlocks some saved-state size savings in D21233334. In order to make use of the normalized file cache to reduce string duplication, we need to have it all happen in the same process. If data is processed in an external process, it gets marshaled and then when the master process unmarshals it, new strings will be allocated anyway.

This increases the time spent creating saved state, but since it helps us shrink the saved state, I think it's worth it. A smaller saved state means we have to store less data, and users have to spend less time waiting for the saved state to download and then load, especially since D21233334 mostly reverses this regression.

Reviewed By: samwgoldman

Differential Revision: D21260927

fbshipit-source-id: b8bc0409ee6605a2c0b6711bfc05de97037eab2e