[parser] disallow newline after type keyword in type aliases
Summary:
Flow's type aliases are designed to squeeze into an unambiguous unused portion of the grammar: `type` is not a reserved word, so `type T` is otherwise two identifiers in a row and that's never valid JS.

however, if there's a line separator between the two, ASI makes it valid JS:

```
type
Foo = {}

// valid JS, as if you'd written:
type;
Foo = {};
```

so, if there's a line separator after `type`, we must interpret it as an expression statement instead of a type alias.

this is also what typescript does.

Fixes https://github.com/facebook/flow/issues/7789

Reviewed By: gkz

Differential Revision: D21328235

fbshipit-source-id: f8564d67cb629228d1524ee4d7b3c7268e3f388c