Add command to drive merge via type sigs instead of sig ASTs
Summary:
This diff adds a new internal-only command `flow type-sig-check`, which can be
used to test the type-sig code against real world repos.

The main contribution here is the `Type_sig_merge` module, which converts
resolved signatures into `Context.sig_t` directly. This is very different from
the existing merge phase in types-first which works over signature ASTs.

Note that producing signature contexts is not the goal of the type-sig project.
Indeed, the resulting sig cxs are not any more compact than the existing ones.
They have the same amount of copied information in them and have the same level
of granularity, components.

Instead, the intermediate goal here is to show the type signatures have enough
information in them to actually produce types in practice.

TODO:
[x] Run successfully over xplat
[x] Run successfully over www
[x] Return type-sig errors from SigParse phase
[x] Replace master_cx with one built using type sigs
[x] Replace all existing types-first infra (sig file sigs) with new infra
[ ] Write a test harness to compare with types-first / sig ASTs
[ ] Add more comments throughout

Reviewed By: panagosg7

Differential Revision: D18680782

fbshipit-source-id: ef69acfca7780fed9ea8b338412c0a7512082fab