quote 'raw' part of Property.Literal in Ty_serializer
Summary:
The Ty_serializer would not replace the quotes in property keys that required
them, leading to parse errors. E.g. `{ '-': 1 }` would be printed as `{ -: 1 }`.
This diff restores them.

Ideally, we would reuse the `raw` property as parsed, and thread it all the way
through inference. This is a lot more invasive that justified for this use case.

Reviewed By: mroch

Differential Revision: D21342332

fbshipit-source-id: 493aa8cd084d55768abe95b7e2d4688826fb7985