[enums] Fix issue when exhaustively checking on invalid member name
Summary:
In D21259822 I changed the executives checking logic to check for the enum object rather than the enum type itself, which fixed a bug. Invalid member access is its own error, but if you do have that error while attempting the exhaustive check, then the `SMap.find member_name seen` would crash, since the invalid member name would not be in the map.

Fix this by checking that the member name being looked at is in the members map. (this check was not necessary before D21259822 since having a valid enum value would require a valid member name).

Reviewed By: dsainati1

Differential Revision: D21365838

fbshipit-source-id: 198c3ccc2b63efa57031441926872adb3d9ba84e