[ty] make literal flag optional
Summary:
The normalizer already had a `preserve_inferred_literal_types` option
that determined if "special" literal types would be inferred. Until now,
this flag only changed the behavior of the normalizer for primitive types.
With this change, we extend this behavior to arrays and objects.

This also makes the `literal` flag for normalized array and object types optional.
It will be `None` if `preserve_inferred_literal_types` is false.

This has a couple of advantages:
* by not setting the literal flag during normalization, we have more opportunities
for better optimization of types, since literal and their respective non-literal
version will be simplified away.
* we can now isolate handling of literals in the normalizer and the printers,
and clean up some relevant hardcoded fixes in autofix and annotate-exports.

Reviewed By: samwgoldman

Differential Revision: D21323881

fbshipit-source-id: f34bc4d2526265e1f3cd7faee6561a216ac30524