Do not wrap single template literal argument when printing call
Summary:
Let's match prettier's behavior when printing call expressions with a single template literal (or tagged template) argument. Like prettier, if we observe that a call expression has a single template literal argument and no comments, we should preserve whether the template literal begins on a new line or not. We can accomplish this by comparing the template's start loc to the argument list's start loc.

```
// Now, the following will be reprinted without changes
foo(`a
b`)

// Where before the template was always being moved to a new line
foo(
  `a
b`
)
```

Reviewed By: gkz

Differential Revision: D21159812

fbshipit-source-id: ec28e7bd990c230c3f5e28b8bc0a3ba34342eaee