Comment AST node stores whether comment starts on new line
Summary:
The majority of modifications made when reprinting files that have not yet been fixed are all of the same class - not preserving newlines before comments.

For example, consider the case of variable declarations or object properties. If there is a leading comment on the initializer or value, newlines before that comment will not be preserved when reprinting. Even worse, after prettier runs the comment may actually be moved to a different node:

```
// Original
const x =
  // L
  1;

{
  key:
    // L
    value,
}

// After reprinting
const x = // L
  1;

{
  key: // L
    value,
}

// After reprinting and prettier
const x = // L
  1;

{
  // L
  key: value,
}
```

To fix this we want to reprint these leading comments to preserve whether they are on a new line or not, so that both of the following can be reprinted and run through prettier without any changes:

```
const x = // L
  1;

const x =
  // L
  1;
```

Unfortunately, we do not have the necessary information to preserve these newlines when reprinting. This is because we do not always keep track of the location of the token before the comment - it would be very hard to find this location generally, and there are also cases where the previous token's loc is never actually added to the AST.

To fix this we can add modify our representation of comments in the AST to include an `on_newline` bool field which tracks whether the comment starts on a new line or not (ignoring leading whitespace on the line). This involved changing the representation of comments in the AST to be a single record with a `kind` variant, which can be `Block` or `Line`, as well as a raw `text` string and the aforementioned `on_newline` bool. This actually simplifies a lot of places where we deal with comments, since we usually want to extract the text regardless of whether the comment is a line or block comment, which no longer requires matching over both `Line` and `Block` variants.

To populate the `on_newline` field we now keep track of the `last_lex_loc` in the lexer's `lex_env`, which is the location of the last lexed comment or token. We must update the `lex_env` with this value every time we lex a comment or token. We then can populate the `on_newline` field for comments by checking whether the previous lex loc was on an earlier line than the loc of the current comment.

This change only modifies the AST and populates the `on_newline` field. The next change will start to fix the comment printing described above.

Reviewed By: mroch

Differential Revision: D21166734

fbshipit-source-id: ceee9d3b5330a2e0556222e24f3afcd12de0052a