Fix printing of comments attached to single simple parameter of arrow function
Summary:
We currently have an issue with reprinting simple arrow function parameters with attached comments. Arrow functions parameters that consist of a single identifier parameter are currently printed without parens. However when reformatted with prettier, prettier will then wrap the single parameter in parens and then places the comments around the parens.

```
// Original
(/*L*/ x /*T*/) => {}

// After reprinting
/*L*/ x /*T*/ => {}

// After prettier
/*L*/ (x) /*T*/ => {}
```

We can fix this by detecting this case in the `js_layout_generator` and printing the arrow function params enclosed in parentheses like normal. More specifically, we only need to print the params enclosed in parentheses if there is a comment on the same line as the param. If all comments appear before the line with the param then parentheses are not needed:

```
// Reprinted without modification
// L
x => {}
```

Reviewed By: gkz

Differential Revision: D21167651

fbshipit-source-id: 6e15b6d5b7e0c6a0b380f987f06f869641e29f8e