Fix printing of comments before the period in member expressions
Summary:
When printing member expressions, block comments directly before the period are not reprinted correctly.

```
// Before
foo()
  /* Comment */
  .bar();

// After reprinting
foo()
  /* Comment */ .bar()
```

Instead, we should detect the case when the last trailing comment of the object of the member expression is a block comment on a line before the property. In this case we should print an extra newline after the block comment so that it will be on the line above the dot and property.

We do not need to add any special handling for line comments, as they are already followed by a hard line break.

Reviewed By: gkz

Differential Revision: D21208869

fbshipit-source-id: 83893bfe733fc7d4a7b5bb1a1841aa192a47cf5f