Fix printing of leading comments within parenthesized expression
Summary:
When printing leading comments within parenthesized expressions we are printing the comments on the same line as the opening paren.

```
// Before
(
  // L
  a
)

// After printing
( // L
  a
)
```

We can detect this case by fetching the comment bounds of the inner expression and checking if there are any leading comments. If so, we should insert a line break after the opening paren so that the line comment is printed on the next line. Let's update the shared parentheses wrapping functions to check for this case.

Reviewed By: pieterv

Differential Revision: D21212142

fbshipit-source-id: ea534747c38949f38c1cb768011f7648825f8b7f