Preserve newline before comments when printing else clauses
Summary:
Leading comments on else clauses are not always being reprinted correctly at the moment. Whether or not the leading comment starts on a newline is not preserved when printing else clauses. We can use the recently created `comment_aware_separator` utility to fix this.

However for accurate reprinting, we need the loc of the entire alternate clause, including both the `else` and the statement body. This was not previously being stored in the AST, so let's update the AST to store this loc in the standard way.

Now we can reprint both of the following without modifications:

```
if (true) {}
// L alternate
else {}
```

Reviewed By: pieterv

Differential Revision: D21167244

fbshipit-source-id: a83ef15fbf88ba3b8e24cf20f6359a2c777c6efa