Fix printing of return arguments with leading comments
Summary:
We are currently missing a very important case when reprinting the argument of a return statement. If the return statement contains any leading comments, we need to wrap the entire argument and its comments in parentheses. Otherwise, it may be printed in such a way that the return statement no longer has an argument.

Consider the case where the return argument has a leading line comment (or a block comment on an earlier line). When printing back the argument we first print the comments, which will appear after the `return` keyword. Then a line break is printed, and then the argument. However, when parsing return statements if there is a newline after the `return` keyword and before the argument the return statement is interpreted as not having an argument. This means that if the return argument has leading comments, our printer can accidentally print it back without an argument:

```
// Before printing
return (
  // L
  arg;
);

// After printing
return // L
  arg;

// How its actually parsed
return; // L
arg;
```

We can fix this by detecting when the return argument has any leading comments using the comment bounds utilities. If there are leading comments, we need to pass the argument through `wrap_in_parens_on_break` so that parentheses will be inserted if there are any line breaks from leading comments.

The same logic applies to the argument of throw statements since a line break after the `throw` keyword is a syntax error, however this is already handled by always wrapping the throw argument in `wrap_in_parens_on_break`.

Reviewed By: gkz

Differential Revision: D21213304

fbshipit-source-id: ae8fb2562f0f9088b144392b6e32c89467cb0c67