Fix printing of trailing line comments in lists
Summary:
When printing lists that have trailing line comments on items, we are running into a bug with our reprinting + prettier.

```
// Before
[
  a // T
  ,
  // L
  b,
]

// After reprinting
[
  a // T
  , // L
  b,
]

// After prettier
[
  a, // T // L
  b,
]
```

To prevent hitting this weird prettier behavior, we need to make sure that a newline is printed after the comma so that the line comment for the next item is printed on a new line. We can get this behavior by adjusting the comment bounds of these items, removing the trailing line comment bound if it appears. This is actually the correct behavior here since a trailing line comment for these list items must have appeared on the same line as the item, and since the comma is printed after it we do not want to follow the usual rule where we don't print a newline after the comma if there is a trailing line comment. This results in the following reprinting, which plays well with prettier:

```
// After reprinting
[
  a // T
  ,
  // L
  b,
]

// After prettier
[
  a, // T
  // L
  b,
]
```

We should remove the trailing line comment bound from the comment bounds of items in comma separated lists (objects, arrays, arguments, params).

Reviewed By: pieterv

Differential Revision: D21210416

fbshipit-source-id: 4179358857f0a2d356e48176923ddfe2f21b0bfa