Do not print extra empty line between leading comments and union/intersection type
Summary:
We currently have a bug with how union/intersection types are printed when they have leading comments. When union/intersection types break across multiple lines, a line break is inserted before the beginning of the union/intersection type. Unfortunately if there was a leading comment on an earlier line, the comment will be printed with a newline at the end so the end result is a blank line appearing between the last leading comment and the beginning of the union/intersection type.

We can fix this by detecting when the last leading comment for the type is on an earlier line. If so, we set the start "inline" value to true, which will cause an extra leading linebreak to not be inserted.

While here let's also use the `comment_aware_separator` utility to handle leading comments on inner types within union/intersections. This will correctly preserve the newline before leading comments on these inner types.

Reviewed By: pieterv

Differential Revision: D21189233

fbshipit-source-id: c1d69485af5b104816da1a73bf8dd4201eecb010