Preserve newlines before comments when printing common nodes
Summary:
The majority of modifications made when reprinting files that have not yet been fixed are all of the same class - not preserving newlines before comments.

For example, consider the case of variable declarations or object properties. If there is a leading comment on the initializer or value, newlines before that comment will not be preserved when reprinting. Even worse, after prettier runs the comment may actually be moved to a different node:

```
// Original
const x =
  // L
  1;

{
  key:
    // L
    value,
}

// After reprinting
const x = // L
  1;

{
  key: // L
    value,
}

// After reprinting and prettier
const x = // L
  1;

{
  // L
  key: value,
}
```

To fix this we want to reprint these leading comments to preserve whether they are on a new line or not, so that both of the following can be reprinted and run through prettier without any changes:

```
const x = // L
  1;

const x =
  // L
  1;
```

We just annotated the AST with whether each comment begins on a new line or not, so we can now use this information to properly print back many of the cases where modifications were being made.

Let's create a new utility function, `comment_aware_separator`, that takes in comment bounds and a separator and returns the separator if there are no leading comments. If there are leading comments and the first comment starts on a new line we should ignore the separator and instead return a hard line. This ensures that comments that begin on a new line in these locations will be reprinted as starting on a new line. If the first comment does not start on a new line then either return an optionally specified separator or the original separator if no `same_line` separator was supplied.

We can then use this utility to fix newlines before comments in the following common locations:

- Variable declaration, between `=` and init expression
- Assignment expression, between `=` and right hand side
- Type alias, between `=` and type
- Arrow function, between `=>` and body
- Object properties, between `:` and value
- Binary and logical expressions, between operator and right hand side

Reviewed By: mroch

Differential Revision: D21166773

fbshipit-source-id: cfe0fcd2a4104ca08a273fdf1d38403175bfd951